requirements.txt
# Semiotic World Agent - Complete Dependencies
# Unified 2D→3D Pipeline with PyQt6 UI
# Consolidated from all requirements files

# =============================================================================
# CORE UI DEPENDENCIES
# =============================================================================
PyQt6==6.9.0
PyQt6-WebEngine==6.9.0
PyQt6_sip==13.10.2

# =============================================================================
# MACHINE LEARNING & AI DEPENDENCIES
# =============================================================================
# PyTorch (with CUDA support)
torch
torchvision
torchaudio

# Hugging Face ecosystem
transformers==4.56.1
diffusers==0.35.1
accelerate==1.10.1
safetensors==0.6.2
huggingface-hub==0.34.4

# =============================================================================
# IMAGE PROCESSING
# =============================================================================
Pillow==11.0.0
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
numpy==2.2.6
scikit-image==0.25.2
scipy==1.16.1

# =============================================================================
# 3D PROCESSING & VISUALIZATION
# =============================================================================
trimesh==4.8.1
open3d==0.19.0
pymeshlab==2025.7

# =============================================================================
# BACKGROUND REMOVAL
# =============================================================================
rembg==2.0.57

# =============================================================================
# IMAGE ENHANCEMENT & UPSCALING
# =============================================================================
basicsr==1.4.2
realesrgan==0.3.0

# =============================================================================
# UTILITIES & CONFIGURATION
# =============================================================================
python-dotenv>=1.0.1
tqdm==4.67.1
requests==2.32.5
psutil==7.0.0
colorama==0.4.6
filelock==3.13.1
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10
urllib3==2.5.0
Brotli==1.1.0
zstandard==0.24.0

# =============================================================================
# JUPYTER & NOTEBOOK SUPPORT
# =============================================================================
papermill>=2.5.0
nbformat>=5.10.4

# =============================================================================
# ADVANCED FEATURES
# =============================================================================
# For advanced image processing
imageio>=2.31.0
matplotlib>=3.7.0
svgwrite>=1.4.3

# For model fine-tuning and optimization
peft>=0.17.1

# For web-based 3D viewer (if needed)
flask>=2.3.0
flask-cors>=6.0.1

# =============================================================================
# CORE DEPENDENCIES
# =============================================================================
# Essential libraries
attrs==25.3.0
Cython==3.1.3
einops==0.8.1
Jinja2==3.1.4
lazy_loader==0.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
omegaconf==2.3.0
onnxruntime==1.22.1
protobuf==6.32.0
PyYAML==6.0.2
regex==2025.9.1
sentencepiece==0.2.1
sympy==1.13.3
timm==1.0.19
tokenizers==0.22.0
typing_extensions==4.12.2
antlr4-python3-runtime==4.9.3

# =============================================================================
# INSTALLATION NOTES
# =============================================================================
# 1. Install PyTorch with CUDA support:
#    pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118
#
# 2. For CPU-only installation:
#    pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cpu
#
# 3. Install all dependencies:
#    pip install -r requirements.txt
#
# 4. Verify installation:
#    python semio_agent_fixed.py